% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_merge.R
\name{clean_merge}
\alias{clean_merge}
\title{clean_merge}
\usage{
clean_merge(redcap, codebook)
}
\arguments{
\item{redcap}{Data from the REDCap instruments}

\item{codebook}{REDCap Codebook containing the adverse event, reporting, and protocol deviation instruments.}
}
\value{
A list containing the date and datetime variables and cleaned REDCap data
}
\description{
Extracts all date and datetime variables and from the data dictionary and cleans REDCap data
}
\details{
Extracts all date and datetime variables and from the data dictionary and cleans REDCap data
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}
 \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{mutate_all}}
 \code{\link[dplyr]{vars}}, \code{\link[dplyr]{any_of}}
 \code{\link[lubridate]{ymd}}
}
