% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinicalTrialsR.R
\docType{package}
\name{ClinicalTrialsR}
\alias{ClinicalTrialsR}
\title{ClinicalTrialsR: R Package for Automating Clinical Trial Reporting.}
\description{
The purpose of this R Package is to automate clinical trial reporting for
  federally funded research studies.
}
\section{ClinicalTrialsR functions}{


\code{\link{get_dsmc_report}}

\code{\link{load_data}}

\code{\link{clean_merge}}
}

