% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dsmc_report.R
\name{get_dsmc_report}
\alias{get_dsmc_report}
\title{get_dsmc_report}
\usage{
get_dsmc_report(datadir = NULL)
}
\arguments{
\item{datadir}{Directory with the input files, Default: NULL}
}
\value{
Returns a Word document containing the DSMC Report
}
\description{
Run the Data Safety Monitoring Committee (DSMC) Report
}
\details{
Run the Data Safety Monitoring Committee (DSMC) Report
}
\seealso{
\code{\link[readr]{read_delim}}
 \code{\link[janitor]{clean_names}}
 \code{\link[rmarkdown]{render}}
}
