% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(codebook)
}
\arguments{
\item{codebook}{REDCap Codebook containing the adverse event, reporting, and protocol deviation instruments.}
}
\value{
Returns a lookup tibble generated from the data dictionary
}
\description{
Creates a lookup table from the data dictionary
}
\details{
Creates a lookup table from the data dictionary
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}
 \code{\link[tidyr]{separate}}, \code{\link[tidyr]{pivot_longer}}
}
