% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_recruitment_timeline.R
\name{set_recruitment_timeline}
\alias{set_recruitment_timeline}
\title{set_recruitment_timeline}
\usage{
set_recruitment_timeline(
  start_date,
  end_date,
  recruited_by,
  n,
  screening_multiple = 2,
  weeks_to_enroll
)
}
\arguments{
\item{start_date}{The clinical trial start date}

\item{end_date}{The clinical trial end date}

\item{recruited_by}{A vector of dates for}

\item{n}{The total number of participants that will be recruited for the clinical trial}

\item{screening_multiple}{A target goal for participant screening that equals `n` multiplied by the `screening_multiple`, Default: 2}

\item{weeks_to_enroll}{The number of weeks prior to the dates expressed in `recruited_by` that is needed to recruit the participants}
}
\value{
Returns a data set and a plot of the recruitment timeline
}
\description{
A function to generate a recruitment timeline that can be
    adjusted manually to set an exact schedule.
}
\details{
A function to generate a recruitment timeline that can be
    adjusted manually to set an exact schedule.
}
\seealso{
\code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate}}
}
